+++ include/lwpa/socket.h
@@ -88,7 +88,7 @@ typedef uint32_t lwpa_poll_events_t;
  *  Refer to the similarly-named option on your favorite man page for more details.
  *  @{ */
 /*! Get/Set, value is int. Value indicates TTL for IPv4 or hop limit for IPv6. */
-#define LWPA_IP_TTL             11 
+#define LWPA_IP_TTL             11
 /*! Get/Set, value is LwpaIpAddr indicating interface IP address for IPv4, or int representing
  *  interface index for IPv6. */
 #define LWPA_IP_MULTICAST_IF    12
@@ -97,7 +97,7 @@ typedef uint32_t lwpa_poll_events_t;
 #define LWPA_IP_MULTICAST_LOOP  14 /*!< Get/Set, value is boolean int */
 /*! [Legacy IPv4-only option, use of #LWPA_MCAST_JOIN_GROUP is preferred] Set only, value is
  *  LwpaMreq. */
-#define LWPA_IP_ADD_MEMBERSHIP  15 
+#define LWPA_IP_ADD_MEMBERSHIP  15
 /*! [Legacy IPv4-only option, use of #LWPA_MCAST_LEAVE_GROUP is preferred] Set only, value is
  *  LwpaMreq. */
 #define LWPA_IP_DROP_MEMBERSHIP 16
@@ -129,9 +129,9 @@ typedef struct LwpaMreq
 /*! Option value for #LWPA_MCAST_JOIN_GROUP and #LWPA_MCAST_LEAVE_GROUP. */
 typedef struct LwpaGroupReq
 {
-  /*! Index of newtork interface on which to join the multicast group. This index is provided in
+  /*! Index of network interface on which to join the multicast group. This index is provided in
    *  LwpaNetintInfo structures. */
-  int ifindex;
+  unsigned int ifindex;
   /*! Multicast group to join. */
   LwpaIpAddr group;
 } LwpaGroupReq;